pyproject.toml
[project]
name = "shortify"
version = "0.1.0"
description = "Shortify is a URL shortener RESTful API built with Python and FastAPI"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "uvicorn",
    "beanie",
    "passlib",
    "python-jose[cryptography]",
    "structlog",
    "pydantic-settings",
    "pydantic[email]",
    "python-multipart",
    "motor>=3.7.1",
    "bcrypt==4.3.0",
    "orjson>=3.11.5",
]

[tool.uv]
dev-dependencies = [
    "ruff",
    "ty",
    "pre-commit",
]

[tool.ruff]
exclude = [
    ".env",
    ".git",
    ".mypy_cache",
    ".pytest_cache",
    ".venv",
]
fix = true
unsafe-fixes = true
show-fixes = true
target-version = "py313"
line-length = 120

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
ignore = [
    "ANN401",
    "D",
    "EM",
    "FA",
    "FBT",
    "G004",
    "INP001",
    "ISC001",
    "PLR",
    "RET501",
    "S311",
    "TC001",
    "TC002",
    "TC003",
    "TRY003",
    "UP040",
    "PERF203",
]
select = ["ALL"]

[tool.ruff.lint.extend-per-file-ignores]
"__init__.py" = ["F401", "F403"]
"tests/*.py" = [
    "ANN001",
    "ANN002",
    "ANN003",
    "ANN401",
    "PLR2004",
    "S101",
    "S311",
]

[tool.ruff.lint.isort]
no-lines-before = ["local-folder", "standard-library"]
known-third-party = []
known-local-folder = []
lines-after-imports = 2
